+++ Qapple/Qapple.xcodeproj/xcshareddata/xcschemes/Qapple.xcscheme
@@ -58,7 +58,7 @@
       <CommandLineArguments>
          <CommandLineArgument
             argument = "-FIRDebugEnabled"
-            isEnabled = "YES">
+            isEnabled = "NO">
          </CommandLineArgument>
       </CommandLineArguments>
    </LaunchAction>
